Move the pncounter tests over to the /test directory
