+++ test/test.rs
@@ -8,3 +8,4 @@ mod lwwreg;
 mod map;
 mod mvreg;
 mod orswot;
+mod pncounter;
